% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ChisqBin.R
\name{chisq.bin}
\alias{chisq.bin}
\title{Chi-square goodness of fit test for binomial distribution}
\usage{
chisq.bin(x, f, n = NULL, p = NULL)
}
\arguments{
\item{x}{The observed values.}

\item{f}{The observed counts.}

\item{n}{Binomial parameter n.}

\item{p}{Binomial parameter pi.}
}
\value{
It returns a table with the contribution to the chi-square statistic for each category,
the chi-square statistic, the degrees of freedom, and the p-value.
}
\description{
This function performs a chi-square goodness of fit test for a binomial distribution.
}
\details{
If \code{p} is not specified, then it is estimated from the data.
If there are categories with expected counts less than 5 or less than 1 a warning
is shown.
}
\examples{
x <- 0:6
f <- c(334, 369, 191, 63, 22, 12, 9)
chisq.bin(x, f, n = 10)
}
\author{
Raul Eyzaguirre.
}

