% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChisqPois.R
\name{chisq.pois}
\alias{chisq.pois}
\title{Chi-square goodness of fit test for Poisson distribution}
\usage{
chisq.pois(x, f, lambda = NULL)
}
\arguments{
\item{x}{The observed values}

\item{f}{The observed counts.}

\item{lambda}{Poisson parameter.}
}
\value{
It returns a table with the contribution to the chi-square statistic for each category,
the chi-square statistic, the degrees of freedom, and the p-value.
}
\description{
This function performs a chi-square goodness of fit test for a Poisson distribution.
}
\details{
If \code{lambda} is not specified, then it is estimated from the data.
If there are categories with expected counts less than 5 or less than 1 a warning
is shown.
}
\examples{
x <- 0:9
f <- c(6, 16, 48, 77, 72, 72, 46, 39, 15, 9)
chisq.pois(x, f)
}
\author{
Raul Eyzaguirre.
}

