% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/EMtdist.R
\name{emtd}
\alias{emtd}
\title{Location and scale parameters estimation of a t distribution}
\usage{
emtd(y, v, initmu = mean(y), inits = sd(y), tol = 1e-04)
}
\arguments{
\item{y}{The data.}

\item{v}{Degrees of freedom.}

\item{initmu}{Initial value for the location parameter.}

\item{inits}{Initial value for the scale parameter.}

\item{tol}{Tolerance for the iterative procedure.}
}
\value{
It returns the estimated location and scale parameters for each iteration.
}
\description{
EM algorithm to estimate the location and scale of a t distribution
for given degrees of freedom.
}
\details{
By default the initial values are set to the sample mean and
standard deviation.
}
\examples{
# Some data
y = c(10,12,16,15,15,17,20,21,16,24,13,22,14,15,16,16,17,18,19,18,23,20,30)

# Estimates for a t(10)
emtd(y, 10)
}
\author{
Raul Eyzaguirre.
}

