% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mdaplot.R
\name{mdaplot}
\alias{mdaplot}
\title{Simulate and plot from a normal distribution}
\usage{
mdaplot()
}
\value{
It returns a histogram and a boxplot for the simulated data.
}
\description{
This function simulates 1000 random samples from a skew normal distribution
for specified values of the mean, standard deviation and skewness parameter.
}
\details{
It uses package \code{sn} to simulate the data and package \code{shiny}
for the web layout. Type mdaplot() in the R console to run the app.
}
\author{
Raul Eyzaguirre.
}

