% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/METdata.R
\docType{data}
\name{METdata}
\alias{METdata}
\title{Yields for genotypes in different environments}
\format{A data frame with 4 columns and 288 rows.}
\source{
International Potato Center, sweetpotato experimental data.
}
\usage{
METdata
}
\description{
This data set gives the yields per plot for 8 genotypes
in 12 environments with a RCBD with 3 blocks in each environment.
}

