% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minota.R
\name{minota}
\alias{minota}
\title{Predice la nota final del curso EP1 y EP2}
\usage{
minota(curso = NULL, vez = NULL, pp = NULL, prob = 0.95, pa1 = NULL,
  pa2 = NULL, pa3 = NULL, pa4 = NULL, pi1 = NULL, pi2 = NULL,
  ep = NULL)
}
\arguments{
\item{curso}{1 o 2 (corresponde a EP1 o EP2).}

\item{vez}{Número de veces que se lleva el curso (1, 2 o 3).}

\item{pp}{Promedio ponderado.}

\item{prob}{Probabilidad para la predicción.}

\item{pa1}{Práctica de aula 1.}

\item{pa2}{Práctica de aula 2.}

\item{pa3}{Práctica de aula 3.}

\item{pa4}{Práctica de aula 4.}

\item{pi1}{Práctica integrada 1.}

\item{pi2}{Práctica integrada 2.}

\item{ep}{Examen parcial.}
}
\value{
Devuelve la nota final estimada con un intervalo de predicción,
y el coeficiente de determinación del modelo.
}
\description{
Esta función predice la nota final del curso basado en datos históricos y un
modelo de regresión lineal.
}
\details{
No es necesario introducir todos los parámetros, el modelo solo
considera los que son introducidos.
}
\examples{
minota(curso = 1, pa1 = 12)
}
\author{
Raúl Eyzaguirre.
}

