% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/MultCoef.R
\name{multcoef}
\alias{multcoef}
\title{Multinomial coefficient}
\usage{
multcoef(n, counts)
}
\arguments{
\item{n}{The size of \eqn{M}.}

\item{counts}{The counts for the repeated elements.}
}
\value{
It returns the multinomial coefficient
\deqn{\frac{n!}{n_1!n_2! \ldots n_k!}}
where
\deqn{n = n_1 + n_2 + \ldots + n_k.}
}
\description{
Computes the number of permutations of a multiset \eqn{M} of size \eqn{n}.
}
\details{
For a set \eqn{M} with \eqn{k} unique elements with associate counts
\eqn{n_1, n_2, \ldots, n_k}, you only need to specify in the \code{counts}
argument the counts that are bigger than 1.
}
\examples{
# The number of permutations of the letters in the set M = {A,A,A,B,B,C}
multcoef(6, c(3,2,1))
# Same result with
multcoef(6, c(3,2))
}
\author{
Raul Eyzaguirre.
}

