% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/MissValEst.R
\name{mve.rcbd}
\alias{mve.rcbd}
\title{Estimation of missing values for a RCBD}
\usage{
mve.rcbd(trait, geno, rep, data, maxp = 0.05, tol = 1e-06)
}
\arguments{
\item{trait}{The trait to estimate missing values.}

\item{geno}{The genotypes.}

\item{rep}{The replications or blocks. A RCBD is assumed.}

\item{data}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, defaults to 5\%.}

\item{tol}{Tolerance for the convergence of the iterative estimation process.}
}
\value{
It returns a data frame with name \code{new.data} and the number
\code{est.num} and proportion \code{est.prop} of estimated missing values.
The \code{new.data} data frame contains the experimental layout and columns
\code{trait} and \code{trait.est} with the original data and the original data plus the
estimated values.
}
\description{
Function to estimate missing values for a Randomized Complete Block Design (RCBD).
}
\details{
A \code{data.frame} with data for a RCBD with at least two replications
and at least one datum for each treatment must be loaded. Experimental data
with only one replication, any treatment without data, or more missing values than
specified in \code{maxp} will generate an error message.
}
\author{
Raul Eyzaguirre.
}

