% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/CheckConsis.R
\name{spconsis}
\alias{spconsis}
\title{Check consistency for sweetpotato experimental data}
\usage{
spconsis(plot.size, data)
}
\arguments{
\item{plot.size}{Plot size in square meters.}

\item{data}{The name of the data frame.}
}
\value{
It returns a file with name checks.txt.
}
\description{
Set of rules to check for consistency of sweetpotato experimental data.
Data labels must be defined as specified in the PROCEDURES FOR THE EVALUATION
AND ANALYSIS OF SWEETPOTATO TRIALS document.
}
\examples{
# The data
 head(pjpz09)
 str(pjpz09)

 # Check the data
 spconsis(4.5, pjpz09)
}
\author{
Raul Eyzaguirre.
}

