% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/CheckConsis.R
\name{spconsis02}
\alias{spconsis02}
\title{Check consistency for sweetpotato experimental data, part 2}
\usage{
spconsis02(data)
}
\arguments{
\item{data}{The name of the data frame.}
}
\value{
It returns all rows with some kind of inconsistency and all rows with outliers
for internal use.
}
\description{
Set of rules to check for consistency of sweetpotato experimental data.
Part 2: NOPE and dependencies.
}
\author{
Raul Eyzaguirre.
}

