% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/CheckConsis.R
\name{spconsis05}
\alias{spconsis05}
\title{Check consistency for sweetpotato experimental data, part 5}
\usage{
spconsis05(data)
}
\arguments{
\item{data}{The name of the data frame.}
}
\value{
It returns all rows with some kind of inconsistency and all rows with outliers
for internal use.
}
\description{
Set of rules to check for consistency of sweetpotato experimental data.
Part 5: NOPR and number of roots.
}
\author{
Raul Eyzaguirre.
}

