% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/CheckConsis.R
\name{spconsis09}
\alias{spconsis09}
\title{Check consistency for sweetpotato experimental data, part 9}
\usage{
spconsis09(data, plot.size)
}
\arguments{
\item{data}{The name of the data frame.}

\item{plot.size}{Plot size in square meters.}
}
\value{
It returns all rows with some kind of inconsistency and all rows with outliers
for internal use.
}
\description{
Set of rules to check for consistency of sweetpotato experimental data.
Part 9: Calculated variables.
}
\author{
Raul Eyzaguirre.
}

