% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/CheckConsis.R
\name{spconsis11}
\alias{spconsis11}
\title{Check consistency for sweetpotato experimental data, part 11}
\usage{
spconsis11(data)
}
\arguments{
\item{data}{The name of the data frame.}
}
\value{
It returns all rows with some kind of inconsistency and all rows with outliers
for internal use.
}
\description{
Set of rules to check for consistency of sweetpotato experimental data.
Part 11: Outliers detection based on interquartile range and values out of range for DM data.
}
\author{
Raul Eyzaguirre.
}

