% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/CheckConsis.R
\name{spconsis13}
\alias{spconsis13}
\title{Check consistency for sweetpotato experimental data, part 13}
\usage{
spconsis13(data)
}
\arguments{
\item{data}{The name of the data frame.}
}
\value{
It returns all rows with some kind of inconsistency and all rows with outliers
for internal use.
}
\description{
Set of rules to check for consistency of sweetpotato experimental data.
Part 13: Outliers detection based on interquartile range and values out of range for lab data.
}
\author{
Raul Eyzaguirre.
}

