% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfreq.R
\name{tfreq}
\alias{tfreq}
\title{Frequency distribution table}
\usage{
tfreq(data, limits = NULL, open = "right")
}
\arguments{
\item{data}{The observations to construct the frequency distribution table.}

\item{limits}{The class limits.}

\item{open}{Where to leave the class limits open, \code{left} or \code{right}.
Defaults to \code{right}.}
}
\value{
It returns a frequency distribution table with columns for class mark,
absolute and relative frequencies, and cumulative absolute and relative frequencies.
}
\description{
Constructs a frequency distribution table for a quantitative variable.
}
\details{
If class limits are not specified, the Sturges' rule is used to calculate the
number of class intervals \eqn{k}:
\deqn{k \approx 1 + 3.3 \log n}
Then, the left limit for the first class interval is set to the minimum value of the data,
the range \eqn{r} is computed and the size of the class intervals is defined by:
\deqn{c \approx \frac{r}{k}}
where \eqn{c} is rounded up with the same number of decimal places as the data.
}
\examples{
# Some random data from a normal population with mean 10 and standard deviation 1
set.seed(1)
datos <- rnorm(100, 10, 1)
# Data with 3 decimal places
datos <- round(datos, 3)
# A summary of the data
summary(datos)
# Frequency table with 6 specified limits
tfreq(datos, c(7, 8, 9, 10, 11, 12, 13))
# Default method
tfreq(datos)
}
\author{
Raúl Eyzaguirre.
}
